Update Remove-ADSIComputer.ps1
Add Recursive parameter to handle 'The directory service can perform the requested operation only on a leaf object'

+++ AdsiPS/Public/Remove-ADSIComputer.ps1
@@ -1,58 +1,68 @@
 function Remove-ADSIComputer
 {
 <#
-	.SYNOPSIS
-		Function to Remove a Computer Account
-	
-	.DESCRIPTION
-		Function to Remove a Computer Account
-	
-	.PARAMETER Identity
-		Specifies the Identity of the Computer.
-	
-		You can provide one of the following:
-			DistinguishedName
-			Guid
-			Name
-			SamAccountName
-			Sid
-	
-	.PARAMETER Credential
-		Specifies the alternative credential to use.
-		By default it will use the current user windows credentials.
-	
-	.PARAMETER DomainName
-		Specifies the alternative Domain.
-		By default it will use the current domain.
-	
-	.EXAMPLE
-		Remove-ADSIComputer TESTSERVER01
-	
-		This command will Remove the account TESTSERVER01
-	
-	.EXAMPLE
-		Remove-ADSIComputer TESTSERVER01 -whatif
-	
-		This command will emulate disabling the account TESTSERVER01
+.SYNOPSIS
+	Function to Remove a Computer Account
 
-	.EXAMPLE
-		Remove-ADSIComputer TESTSERVER01 -credential (Get-Credential)
-	
-		This command will Remove the account TESTSERVER01 using the alternative credential specified
-	
-	.EXAMPLE
-		Remove-ADSIComputer TESTSERVER01 -credential (Get-Credential) -domain LazyWinAdmin.local
-	
-		This command will Remove the account TESTSERVER01 using the alternative credential specified in the domain lazywinadmin.local
-	
-	.NOTES
-		Francois-Xavier.Cat
-		LazyWinAdmin.com
-		@lazywinadm
-		github.com/lazywinadmin
-	
-	.LINK
-		https://msdn.microsoft.com/en-us/library/system.directoryservices.accountmanagement.computerprincipal(v=vs.110).aspx
+.DESCRIPTION
+	Function to Remove a Computer Account
+
+.PARAMETER Identity
+	Specifies the Identity of the Computer.
+
+	You can provide one of the following:
+		DistinguishedName
+		Guid
+		Name
+		SamAccountName
+		Sid
+
+.PARAMETER Credential
+	Specifies the alternative credential to use.
+	By default it will use the current user windows credentials.
+
+.PARAMETER DomainName
+	Specifies the alternative Domain.
+	By default it will use the current domain.
+
+.PARAMETER Recursive
+    Specifies that any child object should be deleted as well
+    Typically you would use this parameter if you get the error "The directory service can perform the requested operation only on a leaf object"
+    when you try to delete the object without the -recursive param
+
+.EXAMPLE
+	Remove-ADSIComputer TESTSERVER01
+
+	This command will Remove the account TESTSERVER01
+
+.EXAMPLE
+	Remove-ADSIComputer TESTSERVER01 -recursive
+
+	This command will Remove the account TESTSERVER01 and all the child leaf
+
+.EXAMPLE
+	Remove-ADSIComputer TESTSERVER01 -whatif
+
+	This command will emulate disabling the account TESTSERVER01
+
+.EXAMPLE
+	Remove-ADSIComputer TESTSERVER01 -credential (Get-Credential)
+
+	This command will Remove the account TESTSERVER01 using the alternative credential specified
+
+.EXAMPLE
+	Remove-ADSIComputer TESTSERVER01 -credential (Get-Credential) -domain LazyWinAdmin.local
+
+	This command will Remove the account TESTSERVER01 using the alternative credential specified in the domain lazywinadmin.local
+
+.NOTES
+	Francois-Xavier.Cat
+	LazyWinAdmin.com
+	@lazywinadm
+	github.com/lazywinadmin
+
+.LINK
+	https://msdn.microsoft.com/en-us/library/system.directoryservices.accountmanagement.computerprincipal(v=vs.110).aspx
 #>
 	[CmdletBinding(SupportsShouldProcess = $true)]
 	PARAM (
@@ -61,7 +71,9 @@ function Remove-ADSIComputer
 		[Alias("RunAs")]
 		[System.Management.Automation.Credential()]
 		$Credential = [System.Management.Automation.PSCredential]::Empty,
-		[String]$DomainName)
+		[String]$DomainName,
+		[switch]$Recursive
+	)
 	
 	BEGIN
 	{
@@ -78,11 +90,26 @@ function Remove-ADSIComputer
 	{
 		TRY
 		{
-			if ($pscmdlet.ShouldProcess("$Identity", "Remove Account"))
+			# Not Recursive
+			if (-not $PSBoundParameters['Recursive'])
 			{
-				$Account = Get-ADSIComputer -Identity $Identity @ContextSplatting
-				$Account.delete()
+				if ($pscmdlet.ShouldProcess("$Identity", "Remove Account"))
+				{
+					$Account = Get-ADSIComputer -Identity $Identity @ContextSplatting
+					$Account.delete()
+				}
 			}
+			
+			# Recursive (if the computer is the parent of one leaf or more)
+			if ($PSBoundParameters['Recursive'])
+			{
+				if ($pscmdlet.ShouldProcess("$Identity", "Remove Account and any child objects"))
+				{
+					$Account = Get-ADSIComputer -Identity $Identity @ContextSplatting
+					$Account.GetUnderlyingObject().deletetree()
+				}
+			}
+			
 		}
 		CATCH
 		{
